requirements.txt
hydra-core
omegaconf
wandb
torch
transformers
datasets
numpy
tqdm
pandas
transformers[torch]